feat: Connect watch face mdule to app module
Added an intent service to the app listening for the ACTION_DATA_UPDATED event
emitted upon the successful weather data sync. The purpose of this service
is packaging weather data into DataItems and sending them to the watch face.

Added code to the watch face for receiving the weather data and displaying
max and min temperatures.

Renamed the watch face package to match the name of the app package,
because otherwise the sync wasn't working.