fix: Force update watch face when selected from the list
This commit fixes the issue with the watch face not being updated when
the user selects it from the list of available watch faces. Basically
what was happening is the watch face was displaying the initial values
of weather condition (0/0/0) until the app was initiating the sync,
sending data to the watch face and triggering its update.
Now the watch face, when selected, sends a message to the app,
which forces the app to send back the most recent weather conditions.

How it works:
- when the watch face becomes visible, it attempts to establish
a Google API client connection;
- when the client is connected, the watch face sends a specific message
through the Message API (see Engine.requestWeatherUpdate());
- the app has an implementation of WearableListenerService which listens
for the incoming message (see MyWearableListenerService);
- when the listener service receives a message from the watch face,
it starts the service which sends weather data to the watch face.

Notes:
- apparently it's really important to make sure the DataItem messages
are unique, otherwise they are discarded without synchronizing
to the remote devices. See WearableIntentService.onConnected() for
the comments on how to handle this issue.